+++ pyproject.toml
@@ -3,9 +3,9 @@ requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "azure-agent-mcp-adapter"
+name = "azure-agent-mcp-server"
 version = "0.1.0"
-description = "Azure AI Agent Service MCP Adapter"
+description = "Azure AI Agent Service MCP Server"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
@@ -13,4 +13,5 @@ dependencies = [
     "azure-identity~=1.21.0",
     "python-dotenv~=1.0.1",
     "azure-ai-projects~=1.0.0b7",
+    "aiohttp>=3.11.14",
 ]
