--- src/python/azure_agent_mcp_adapter/__init__.py
@@ -1,3 +1,0 @@
-"""Azure AI Agent Service MCP Adapter."""
-
-__version__ = "0.1.0"
