+++ src/python/azure_agent_mcp_server/__init__.py
@@ -1,0 +1,3 @@
+"""Azure AI Agent Service MCP Server."""
+
+__version__ = "0.1.0"
